requirements/requirements.txt
scikit-learn>=1.1.3,<2.0.0
pandas>2.0.0,<3.0.0
scipy>1.0.0,<2.0.0
pyyaml>6.0.0,<7.0.0
tensorflow>2.0.0,<3.0.0
uvicorn<1.0.0
fastapi<1.0.0